requirement.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
ConfigArgParse==1.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.2
django-debug-toolbar==3.7.0
django-filter==22.1
django-redis==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt
djoser
drf-spectacular==0.24.2
exceptiongroup==1.0.0
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.1
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.0
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
oauthlib==3.2.2
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
psutil==5.9.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
pytest-django==4.5.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.5
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
roundrobin==0.0.4
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.5.2
